requirements.txt
mitmproxy
pyyaml
pyfiglet